Added Layers as well as shallow copying/selection
The Root Layer has been added - support for Layers and objects to come
soon.

Changes of drawn shapes is currently possible, although really finnicky.
It requires the user to select a very precise point.

Copying of Descriptions is possible, however any changes to shapes from
the original will update in the original too (just how Shared Pointers
work). This will be worked on soon.

Editing of Definitions is also possible - it will update the description
as well as the name of the definition.