Added Layers, and a program window
Layers, as well as Objects, can now be added to the Program. As a
conseuence, I've also added the ability to run the Program to see what
is rendered. This is currently done in such a way that everything
overlaps, which is not the final intention, but works for now.

Positioning does not work for Objects in Layers (as above), neither do
any Transformations (Rotation, Scaling). Which need to be worked out --
probably change to using a QImage for each Shape, as this should make it
easier. As none of this works, the Textual Portion also does not work
(and is not currently implemented in any way).