Moving shapes to 'real zero' is now possible
The real zero is the bottom left of the screen, as opposed to the top
left of the screen. This, of course, involves moving objects to the
bottom left of the screen. At the moment, each object is placed in it's
correct position near the bottom left of the screen, and each subsequent
object will be placed 200 pixels to the right of the previous object.

This still needs to be worked on, as it doesn't place the object in
exactly the correct position, however it's pretty close.